pyproject.toml
[tool.poetry]
name = "spotifyinternshipnotifier"
version = "0.1.0"
description = "Simple script which notifies you as soon as Spotify posted any internship listings."
authors = ["Franzi <92818644+franziskavonalbedyll@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
azure-functions = "^1.17.0"
azure-communication-email = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
